Significant performance improvement implemented.
Takes advantage of inode info in UNIX. In crawl(), each of the folders in the base start
location are checked to see if they have changed since last we checked and only these are crawled.
To support this, the time of last check in UNIX epoch time is put at the top of the record.txt file.
This change has resulted in an enormous decrease in time spent on disk reads (and minor computation).
Also cleaned up the logic of crawl() calls a bit, but that code block is still very messy.